requirements.txt
fastapi
uvicorn
pdfminer.six
scikit-learn
sentence-transformers
spacy
numpy
